temporary project settings, semi-functional narration style
-added placeholder pitch drop logo splash image
-changed stretch modes and default viewport sizes
-basic "nvl" narration style for narrator character(s) that fits the default viewport
-added a test character that uses premade speaker textbox style

known issues:
"[Dialogic] The glossary file "res://timelines/New_File.tres" is missing. Make sure it exists."
uncertain how to make it exist

"Cannot call method 'call' on a null value."
Uncertain what it is saying this in reference to.

And also some apparently missing sounds that... do seem to be there so I'm confused?